+++ index.html
@@ -10,7 +10,6 @@
     <link rel="stylesheet" href="./pages/components/footer.css" />
     <script type="module" src="./script/import.js"></script>
     <script type="module" src="./script/index.js"></script>
-    <script type="module" src="./script/search.js"></script>
   </head>
   <body>
     <header></header>
