+++ pages/components/header.html
@@ -1,5 +1,5 @@
 <div class="container">
-  <h1 class="logo">EYESMOVIE</h1>
+  <h1 class="logo"><a href="/index.html">EYESMOVIE</a></h1>
   <form id="search">
     <input id="searchBar" type="text" placeholder="검색어를 입력하세요" />
     <button id="searchBtn" type="submit">검색</button>
